수정

+++ src/main/java/com/synclife/roomreservation/AppRepository.java
@@ -16,10 +16,9 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class AppRepository {
 
-//    private final AppMapper appMapper;
     private final JdbcTemplate jdbcTemplate;
 
-    public void createRoom(String roomName, long userId, String location, int capacity) {
+    public void createRoom(String roomName, int userId, String location, int capacity) {
         String insertRoomSql = "INSERT INTO rooms(room_name, user_id, location, capacity) " +
                 "VALUES (?,?,?,?)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
@@ -27,13 +26,11 @@ public class AppRepository {
         jdbcTemplate.update(connection -> {
             PreparedStatement ps = connection.prepareStatement(insertRoomSql, Statement.RETURN_GENERATED_KEYS);
             ps.setString(1, roomName);
-            ps.setLong(2, userId);
+            ps.setInt(2, userId);
             ps.setString(3, location);
             ps.setInt(4, capacity);
             return ps;
         }, keyHolder);
-
-
     }
 
     public List<Map<String, Object>> findAllRooms() {
@@ -42,11 +39,11 @@ public class AppRepository {
 
     public List<Map<String, Object>> findReservationByDate(LocalDateTime dayStart, LocalDateTime nextDayStart) {
         String sql = """
-        SELECT reservation_id, room_id, user_id, start_at, end_at, status
-        FROM reservations
-        WHERE start_at < ? AND end_at > ?
-        ORDER BY room_id, start_at
-    """;
+                    SELECT reservation_id, room_id, user_id, start_at, end_at, status
+                    FROM reservations
+                    WHERE start_at < ? AND end_at > ?
+                    ORDER BY room_id, start_at
+                """;
 
         return jdbcTemplate.queryForList(
                 sql,
@@ -68,11 +65,11 @@ public class AppRepository {
         );
 
         String checkSql = """
-        SELECT reservation_id FROM reservations
-        WHERE room_id=? AND status IN ('BOOKED')
-          AND start_at < ? AND end_at > ?
-        FOR UPDATE
-    """;
+                    SELECT reservation_id FROM reservations
+                    WHERE room_id=? AND status IN ('BOOKED')
+                      AND start_at < ? AND end_at > ?
+                    FOR UPDATE
+                """;
 
         boolean overlap = jdbcTemplate.query(
                 (Connection con) -> {
@@ -80,7 +77,7 @@ public class AppRepository {
                     ps.setLong(1, roomId);
                     ps.setTimestamp(2, Timestamp.valueOf(end));
                     ps.setTimestamp(3, Timestamp.valueOf(start));
-                    return  ps;
+                    return ps;
                 },
                 (ResultSet rs) -> rs.next()
         );
@@ -90,9 +87,9 @@ public class AppRepository {
         }
 
         String insertReservationSql = """
-        INSERT INTO reservations(user_id, room_id, start_at, end_at, status)
-        VALUES (?,?,?,?, 'BOOKED')
-    """;
+                    INSERT INTO reservations(user_id, room_id, start_at, end_at, status)
+                    VALUES (?,?,?,?, 'BOOKED')
+                """;
 
         GeneratedKeyHolder kh = new GeneratedKeyHolder();
 
@@ -109,12 +106,12 @@ public class AppRepository {
     public int cancelReservation(int reservationId) {
 
         String sql = """
-            UPDATE reservations
-            SET status = 'CANCELED'
-            WHERE reservation_id = ?
-              AND status = 'BOOKED'
-              AND start_at > NOW()
-            """;
+                UPDATE reservations
+                SET status = 'CANCELED'
+                WHERE reservation_id = ?
+                  AND status = 'BOOKED'
+                  AND start_at > NOW()
+                """;
 
         return jdbcTemplate.update(con -> {
             PreparedStatement ps = con.prepareStatement(sql);
@@ -125,11 +122,11 @@ public class AppRepository {
 
     public Long findRoomOwnerId(int reservationId) {
         String sql = """
-        SELECT rm.owner_id
-        FROM reservations rv
-        JOIN rooms rm ON rv.room_id = rm.room_id
-        WHERE rv.reservation_id = ?
-        """;
+                SELECT rm.owner_id
+                FROM reservations rv
+                JOIN rooms rm ON rv.room_id = rm.room_id
+                WHERE rv.reservation_id = ?
+                """;
         List<Long> result = jdbcTemplate.query(
                 sql,
                 ps -> ps.setInt(1, reservationId),
